Automated critical point search
Critical_zoom_notebook will find the (pseudo-)critical temperature for a given value of chemical potential (and other parameters.

critical_point_finder finds the critical point for given lambda1 and quark  mass. LONG RUN TIME

Changes:
Looks for minimum of gradient of sigma instead of sigma^(1/3), which is more reliable, and will not accidentally say the transition temperature occurs when sigma goes to zero.

Because of this, it is no longer necessary to remove the zero values of sigma, which allows the maximum value of Temperature to be arbitrarily large.

This makes the code more robust, and allows for a more accurate determination of the critical temperature.

Also removes some spurious points at small temperature values, which means the minimum temperature and max sigma values don't need to be tuned so precisely.