Remove redundancy, check negative mu
critical_point_finder saves the mu, Tc, and order in a list. It checks to see if the current value of mu has already been tested before running the code, to remove redundant calculations.

Also, did some checking for negative mu values, coming to the conclusion that this is redundant, as the results are symmetric for negative mu.